requirements/prod.txt
fastapi>=0.115.12
python-dotenv>=1.0.1
python-telegram-bot>=21.6
python-telegram-bot[job-queue]
SQLAlchemy>=2.0.35